pyproject.toml
[tool.poetry]
name = "events"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
pip = "^23.3.2"
install = "^1.3.5"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"
pillow = "^10.2.0"
django-filter = "^23.5"
channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.2.0"
psycopg = "^3.1.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
